add command date & time

+++ src/commands/echo.h
@@ -7,7 +7,7 @@
 using namespace std;
 
 int echo(vector<string> i){
-    for(int j=1;j<i.size();j++){
+    for(size_t j=1;j<i.size();j++){
         cout << i.at(j) << " ";
     }
     return 0;
